[IMPROVEMENTS] Add JWT for Authentication [FE]
