+++ frontend/src/api/lessonApi.js
@@ -1,10 +1,13 @@
 import API from "./baseApi";
 
+const accessToken = localStorage.getItem("accessToken");
+
 const lessonApi = {
   getLessons: (payload) => {
     const options = {
       method: "GET",
       url: "/lessons",
+      headers: { Authorization: accessToken },
       params: { ...payload },
     };
