+++ frontend/src/api/wordApi.js
@@ -1,23 +1,17 @@
 import API from "./baseApi";
 
+const accessToken = localStorage.getItem("accessToken");
+
 const wordApi = {
   getWordsLearned: (payload) => {
     const options = {
       method: "GET",
       url: "/words/user",
+      headers: { Authorization: accessToken },
       params: { ...payload },
     };
 
     return API.request(options);
-  },
-  postAnswer: async (payload) => {
-    const options = {
-      method: "POST",
-      url: "/results",
-      data: { ...payload },
-    };
-
-    return await API.request(options);
   },
 };
