+++ frontend/src/pages/user/home/Results/hooks/useResults.js
@@ -8,8 +8,8 @@ import { authenticate } from "../../../../../helpers/auth";
 import { setLoading, setCurrentLesson } from "../../../../../store/lessonSlice";
 
 export const useResults = () => {
-  const navigate = useNavigate();
   const dispatch = useDispatch();
+  const navigate = useNavigate();
   const { user } = useSelector((state) => state.currentUser);
   const { lessons, currentLesson } = useSelector((state) => state.lesson);
 
@@ -35,7 +35,6 @@ export const useResults = () => {
         message.error(data.meta.message);
       }
     );
-
     dispatch(setLoading(false));
   }, [navigate, dispatch, user.id, user.user_type, currentLesson]);
 
@@ -45,10 +44,10 @@ export const useResults = () => {
   };
 
   const getNextLesson = () => {
-    const lessonIndex = lessons.findIndex(
+    const lessonIndex = lessons?.lessons.findIndex(
       (lesson) => lesson.id === currentLesson.id
     );
-    return lessons[(lessonIndex + 1) % lessons.length];
+    return lessons?.lessons[(lessonIndex + 1) % lessons?.lessons.length];
   };
 
   const toNextLesson = () => {
